# mobile/swagger.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "",
        "title": "",
        "contact": {},
        "version": ""
    },
    "host": "",
    "basePath": "",
    "paths": {
...
        "/suppliers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取供应商列表，支持按名称、编码、类型和状态筛选，并支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "供应商管理"
                ],
                "summary": "获取供应商列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "供应商名称（模糊查询）",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "供应商编码（模糊查询）",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "manufacturer",
                            "distributor",
                            "retailer",
                            "wholesaler",
                            "other"
                        ],
                        "type": "string",
                        "description": "供应商类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页数量，默认10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取供应商列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Supplier"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的供应商，需提供必要的供应商信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "供应商管理"
                ],
                "summary": "创建供应商",
                "parameters": [
                    {
                        "description": "供应商信息",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/models.Supplier"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或供应商编码已存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/suppliers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取供应商的详细信息，包括基本信息、联系方式、评级等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "供应商管理"
                ],
                "summary": "获取供应商详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "供应商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取供应商详情",
                        "schema": {
                            "$ref": "#/definitions/models.Supplier"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "供应商不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新现有供应商的信息，需提供完整的供应商信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "供应商管理"
                ],
                "summary": "更新供应商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "供应商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "供应商信息",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/models.Supplier"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或供应商编码已被其他供应商使用",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "供应商不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定ID的供应商，如果供应商已被采购单引用则无法删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "供应商管理"
                ],
                "summary": "删除供应商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "供应商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "供应商已被采购单引用，无法删除",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "供应商不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.CreateSupplierRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "type"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "contact_person": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "delivery_terms": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "payment_terms": {
                    "type": "string"
                },
                "qualification": {
                    "type": "string"
                },
                "rating": {
                    "$ref": "#/definitions/models.SupplierRating"
                },
                "status": {
                    "type": "boolean"
                },
                "type": {
                    "$ref": "#/definitions/models.SupplierType"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据"
                },
                "error": {
                    "description": "错误信息，仅在发生错误时返回",
                    "type": "string",
                    "example": "错误信息"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "description": "详细错误信息，仅在开发环境返回",
                    "type": "string"
                },
                "message": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "数据项列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页记录数",
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "models.Supplier": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "city": {
                    "description": "城市",
                    "type": "string"
                },
                "code": {
                    "description": "供应商编码",
                    "type": "string"
                },
                "contactPerson": {
                    "description": "联系人",
                    "type": "string"
                },
                "contactPhone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deliveryTerms": {
                    "description": "交货条件",
                    "type": "string"
                },
                "email": {
                    "description": "电子邮箱",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "供应商名称",
                    "type": "string"
                },
                "note": {
                    "description": "备注",
                    "type": "string"
                },
                "paymentTerms": {
                    "description": "付款条件",
                    "type": "string"
                },
                "qualification": {
                    "description": "资质证明",
                    "type": "string"
                },
                "rating": {
                    "description": "供应商评级",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SupplierRating"
                        }
                    ]
                },
                "status": {
                    "description": "状态：true-启用，false-禁用",
                    "type": "boolean"
                },
                "type": {
                    "description": "供应商类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SupplierType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SupplierRating": {
            "type": "string",
            "enum": [
                "S",
                "A",
                "B",
                "C",
                "D"
            ],
            "x-enum-comments": {
                "SupplierA": "良好",
                "SupplierB": "一般",
                "SupplierC": "较差",
                "SupplierD": "不合格",
                "SupplierS": "优秀"
            },
            "x-enum-varnames": [
                "SupplierS",
                "SupplierA",
                "SupplierB",
                "SupplierC",
                "SupplierD"
            ]
        },
        "models.SupplierType": {
            "type": "string",
            "enum": [
                "manufacturer",
                "wholesaler",
                "agent",
                "other"
            ],
            "x-enum-comments": {
                "SupplierAgent": "代理商",
                "SupplierManufacturer": "生产厂商",
                "SupplierOther": "其他",
                "SupplierWholesaler": "批发商"
            },
            "x-enum-varnames": [
                "SupplierManufacturer",
                "SupplierWholesaler",
                "SupplierAgent",
                "SupplierOther"
            ]
        }
    }
}